add route handler to edit a venue specified by its id

+++ backend/db/migrations/20230119001031-create-venues.js
@@ -29,10 +29,10 @@ module.exports = {
         type: Sequelize.STRING
       },
       lat: {
-        type: Sequelize.INTEGER
+        type: Sequelize.REAL
       },
       lng: {
-        type: Sequelize.INTEGER
+        type: Sequelize.REAL
       },
       createdAt: {
         allowNull: false,
